Test changing the brand of a material
This is a bit involved because originally it was always only testing for materials under the header 'Custom'. Now it needs to also look under the header named after me.

This currently still fails because the material is down under the limit of the window. We need to scroll down. Let's see if the internet has a solution for that...