Force Squish to save the file in the testdata\ directory
It seems that it is no longer possible to automatically save to the custom directory when the latter is set in the AUT subpage of the suite_Cura settings. Normally, if you set there a custom AUT working directory (e.g., shared\testdata) the AUT should run from there, meaning that whenever an open/save file dialog is popping up it should already be in htat custom directory. So, after only adding the file name and pressing Save the file should be saved in that path. Since this is no longer the case though, with this commit we force the file to be saved in the testdata\ directory, because there are some test steps that fail to read the file because there's nothing there.