Added integration with Excel for performance results
Until now the performance results were stored in a .json file. This is not so practical when creating performance benchmarks, so it was decided that we use Excel to store the results from now on.
This commit implements that integration with Excel, but also leaves the 'old way of doing things' there as well, in case it's needed again in the future. Thus, the script first checks which extension
is given to the Performance file and acts accordingly. Below are the most important things this script now does:
- Distinguishes between .xlsx and .json files;
- Adds the data as dataframes to the Excel sheet. If data are already written to the file (e.g., from previous performance runs) it appends the new data in cells under the existing ones. If not, then it also creates the column headers for each metric
- function append_df_to_excel(). This is a new function that will append a DataFrame [df] to a specific sheet inside an existing Excel workbook. If the workbook doesn't exist, then this function will create it.

There are some extra python packages needed in order for Squish to run:
- pandas
- openpyxl
- numpy 1.19.3
While installing these three packages, pip will install also others that are required for these three to operate.